requirements.txt
datasets
tqdm
openai
anthropic
python-dotenv